+++ DSCResources/WindowsDnsServer/WindowsDnsServer.schema.psm1
@@ -96,6 +96,7 @@ Configuration WindowsDnsServer
     . "$resourcePath\windows.Script.skip.ps1"
     . "$resourcePath\windows.Script.RootHint.ps1"
 
-    Import-DscResource -ModuleName xWinEventLog -ModuleVersion 1.2.0.0
-    . "$resourcePath\windows.xWinEventLog.ps1"
+    Import-DscResource -ModuleName ComputerManagementDsc -ModuleVersion 6.2.0.0
+    . "$resourcePath\windows.WindowsEventLog.ps1"
+
 }